ml-service/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
scikit-learn==1.5.1
pandas==2.2.2
numpy==1.26.4
supabase==2.5.0
python-dotenv==1.0.1
scipy==1.13.1
